requirements.txt
python-dotenv~=0.19.0
requests~=2.28.0
lemon~=1.0.2